Removes circular dependency in logs.py &  Adresses comments
